# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Letterpress — a typewriter letter writing app
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface DeleteResponse {
  success: boolean;
}

export interface ShareResponse {
  token: string;
  url: string;
}

export interface LetterStats {
  total: number;
  drafts: number;
  sent: number;
  completed: number;
}

export type LetterStatus = (typeof LetterStatus)[keyof typeof LetterStatus];

export const LetterStatus = {
  draft: "draft",
  completed: "completed",
  sent: "sent",
} as const;

export interface Letter {
  id: number;
  title: string;
  content: string;
  recipientName?: string | null;
  recipientEmail?: string | null;
  senderName?: string | null;
  /** Font family name */
  font?: string | null;
  /** Hex color for ink */
  inkColor?: string | null;
  /** Paper texture identifier */
  paperTexture?: string | null;
  /** Envelope style identifier */
  envelopeStyle?: string | null;
  /** Hex color for wax seal */
  waxSealColor?: string | null;
  /** Symbol/stamp identifier for wax seal */
  waxSealSymbol?: string | null;
  /** Optional embedded image URL */
  imageUrl?: string | null;
  status: LetterStatus;
  shareToken?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type LetterInputStatus =
  (typeof LetterInputStatus)[keyof typeof LetterInputStatus];

export const LetterInputStatus = {
  draft: "draft",
  completed: "completed",
  sent: "sent",
} as const;

export interface LetterInput {
  title: string;
  content: string;
  recipientName?: string;
  recipientEmail?: string;
  senderName?: string;
  font?: string;
  inkColor?: string;
  paperTexture?: string;
  envelopeStyle?: string;
  waxSealColor?: string;
  waxSealSymbol?: string;
  imageUrl?: string;
  status?: LetterInputStatus;
}

export type LetterUpdateInputStatus =
  (typeof LetterUpdateInputStatus)[keyof typeof LetterUpdateInputStatus];

export const LetterUpdateInputStatus = {
  draft: "draft",
  completed: "completed",
  sent: "sent",
} as const;

export interface LetterUpdateInput {
  title?: string;
  content?: string;
  recipientName?: string;
  recipientEmail?: string;
  senderName?: string;
  font?: string;
  inkColor?: string;
  paperTexture?: string;
  envelopeStyle?: string;
  waxSealColor?: string;
  waxSealSymbol?: string;
  imageUrl?: string;
  status?: LetterUpdateInputStatus;
}