# lib/api-zod/src/generated/types/letter.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Letterpress — a typewriter letter writing app
 * OpenAPI spec version: 0.1.0
 */
import type { LetterStatus } from "./letterStatus";

export interface Letter {
  id: number;
  title: string;
  content: string;
  recipientName?: string | null;
  recipientEmail?: string | null;
  senderName?: string | null;
  /** Font family name */
  font?: string | null;
  /** Hex color for ink */
  inkColor?: string | null;
  /** Paper texture identifier */
  paperTexture?: string | null;
  /** Envelope style identifier */
  envelopeStyle?: string | null;
  /** Hex color for wax seal */
  waxSealColor?: string | null;
  /** Symbol/stamp identifier for wax seal */
  waxSealSymbol?: string | null;
  /** Optional embedded image URL */
  imageUrl?: string | null;
  status: LetterStatus;
  shareToken?: string | null;
  createdAt: Date;
  updatedAt: Date;
}