# lib/api-zod/src/generated/types/letterInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Letterpress — a typewriter letter writing app
 * OpenAPI spec version: 0.1.0
 */
import type { LetterInputStatus } from "./letterInputStatus";

export interface LetterInput {
  title: string;
  content: string;
  recipientName?: string;
  recipientEmail?: string;
  senderName?: string;
  font?: string;
  inkColor?: string;
  paperTexture?: string;
  envelopeStyle?: string;
  waxSealColor?: string;
  waxSealSymbol?: string;
  imageUrl?: string;
  status?: LetterInputStatus;
}